feat: add setup command and move logic from main script into command execution function
Setup command provides the same functionallity as previous versions before Cobra. It is the entrypoint for configuring Tailscale on a card.
