+++ main.go
@@ -1,7 +1,3 @@
-/*
-Copyright © 2025 NAME HERE <EMAIL ADDRESS>
-
-*/
 package main
 
 import "github.com/borisdvlpr/gotail/cmd"
